Add hosting diagnostics and refactor builder methods
Introduced new diagnostics (`NALIX040` to `NALIX045`) for
`NetworkApplicationBuilder` to catch common misconfigurations
like missing `BufferPoolManager` or `ConnectionHub` setups,
invalid handler/metadata provider types, and improper TCP/UDP
binding configurations.

Refactored `NalixUsageAnalyzer` to include hosting-related
analysis and removed redundant code. Updated `INetworkApplicationBuilder`
to replace `UseBufferPoolManager` with `ConfigureBufferPoolManager`
for better naming consistency. Adjusted `NetworkApplicationBuilder`
and example code accordingly.

Improved documentation and comments to clarify the purpose
and usage of the new diagnostics and builder methods.
